requirements.txt
Django==1.10.3
django-environ==0.4.0
psycopg2==2.6.2
django-allauth==0.32.0

